Write the skeleton for the analysis module
